1、解决了负数显示错误的问题； 2、解决了设置最大限位值设置无效的问题；

+++ BLEMaster/src/com/gemhu/blemaster/DataPackage.java
@@ -159,12 +159,12 @@ public class DataPackage {
 	}
 	
 	public float getSpeed() {
-		int speed = ((this.mData[3] & 0xFF) << 8 | (this.mData[4] & 0xFF));
+		short speed = (short) ((this.mData[3] & 0xFF) << 8 | (this.mData[4] & 0xFF));
 		return speed * 0.1f;
 	}
 	
 	public float getPos() {
-		int pos = ((this.mData[3] & 0xFF) << 8 | (this.mData[4] & 0xFF));
+		short pos = (short) ((this.mData[3] & 0xFF) << 8 | (this.mData[4] & 0xFF));
 		return pos * 0.1f;
 	}
